+++ file/retrytransfers_test.go
@@ -8,6 +8,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/Files-com/files-sdk-go/v2/lib/timer"
+
 	files_sdk "github.com/Files-com/files-sdk-go/v2"
 	"github.com/Files-com/files-sdk-go/v2/file/manager"
 	"github.com/Files-com/files-sdk-go/v2/file/status"
@@ -19,96 +21,81 @@ import (
 func TestRetryTransfers(t *testing.T) {
 	assert := assert.New(t)
 	buildDownloadTest(func(job *status.Job) {
-		job.RetryPolicy = string(RetryAll)
 		events := status.EventsReporter{}
 		var retryingEvents []status.Status
 		events[status.Retrying] = func(file status.File) {
 			retryingEvents = append(retryingEvents, file.Status)
 		}
 		job.SetEventsReporter(events)
-		RetryTransfers(context.Background(), job)
+		RetryByPolicy(context.Background(), job, RetryAll)
 		assert.Equal(true, job.All(status.Complete))
 		assert.Equal(3, len(retryingEvents), "sets a retrying status before starting")
 	})
 
 	buildDownloadTest(func(job *status.Job) {
-		job.RetryPolicy = string(RetryErrored)
-		RetryTransfers(context.Background(), job)
+		RetryByPolicy(context.Background(), job, RetryUnfinished)
 		assert.Equal(false, job.All(status.Complete))
 		assert.Equal(2, job.Count(status.Complete))
 		assert.Equal(1, job.Count(status.Queued))
 	})
 
 	buildDownloadTest(func(job *status.Job) {
-		job.RetryPolicy = string(RetryErrored)
-		job.Stopped = true
-		RetryTransfers(context.Background(), job)
-		assert.Equal(1, job.Count(status.Complete))
-		assert.Equal(1, job.Count(status.Queued))
-		assert.Equal(1, job.Count(status.Errored), "job was stopped so no retrying")
-	})
-
-	buildDownloadTest(func(job *status.Job) {
-		job.RetryPolicy = string(RetryErroredIfSomeCompleted)
-		RetryTransfers(context.Background(), job)
+		RetryByPolicy(context.Background(), job, RetryErroredIfSomeCompleted)
 		assert.Equal(false, job.All(status.Complete))
 		assert.Equal(2, job.Count(status.Complete))
 		assert.Equal(1, job.Count(status.Queued))
 	})
 
 	buildDownloadTest(func(job *status.Job) {
-		job.RetryPolicy = string(RetryErroredIfSomeCompleted)
 		job.Statuses[0].(*DownloadStatus).lastByte = time.Time{}
 		job.Statuses[1].(*DownloadStatus).lastByte = time.Time{}
 		job.Statuses[2].(*DownloadStatus).lastByte = time.Time{}
-		RetryTransfers(context.Background(), job)
+		RetryByPolicy(context.Background(), job, RetryErroredIfSomeCompleted)
 		assert.Equal(1, job.Count(status.Complete))
 		assert.Equal(1, job.Count(status.Queued))
 		assert.Equal(1, job.Count(status.Errored), "is not retried because completed happened before this error not after")
 	})
 
 	buildUploadTest(func(job *status.Job, _ *MockUploader) {
-		job.RetryPolicy = string(RetryAll)
-		RetryTransfers(context.Background(), job)
+		RetryByPolicy(context.Background(), job, RetryAll)
 		assert.Equal(true, job.All(status.Complete))
 	})
 
 	buildUploadTest(func(job *status.Job, _ *MockUploader) {
-		job.RetryPolicy = string(RetryErrored)
-		RetryTransfers(context.Background(), job)
+		RetryByPolicy(context.Background(), job, RetryUnfinished)
 		assert.Equal(false, job.All(status.Complete))
 		assert.Equal(2, job.Count(status.Complete))
 		assert.Equal(1, job.Count(status.Queued))
-		assert.Equal(false, job.EndTime.IsZero(), "does change the original time")
+		assert.Equal(false, job.Running(), "does change the original time")
 	})
 
 	buildUploadTest(func(job *status.Job, _ *MockUploader) {
-		job.StartTime = time.Now().AddDate(0, -1, 0)
-		startTime := job.StartTime
-		job.EndTime = time.Now().AddDate(0, -1, 0)
-		endTime := job.EndTime
-		job.RetryPolicy = string(RetryErrored)
+		startTime := time.Now().AddDate(0, -1, 0)
+		endTime := time.Now().AddDate(0, -1, 0)
+		job.Timer.Runs = timer.Runs{timer.Run{
+			Start:  startTime,
+			Finish: endTime,
+		}}
 		uploadStatusErrored := job.Statuses[0].(*UploadStatus)
 		uploadStatusErrored.Status = status.Complete
-		RetryTransfers(context.Background(), job)
+		RetryByPolicy(context.Background(), job, RetryUnfinished)
 		assert.Equal(false, job.All(status.Complete))
 		assert.Equal(2, job.Count(status.Complete))
 		assert.Equal(1, job.Count(status.Queued))
-		assert.Equal(startTime, job.StartTime, "does not change the original time")
-		assert.Equal(endTime, job.EndTime, "does not change the original time")
+		assert.Equal(startTime, job.StartTime(), "does not change the original time")
+		assert.Equal(endTime, job.FinishTime(), "does not change the original time")
 	})
 
 	buildUploadTest(func(job *status.Job, _ *MockUploader) {
-		job.RetryPolicy = string(RetryErroredIfSomeCompleted)
 		uploadStatusErrored := job.Statuses[0].(*UploadStatus)
 		uploadStatusComplete := job.Statuses[1].(*UploadStatus)
 		uploadStatusComplete.lastByte = uploadStatusErrored.lastByte
-		RetryTransfers(context.Background(), job)
+		RetryByPolicy(context.Background(), job, RetryErroredIfSomeCompleted)
 		assert.Equal(false, job.All(status.Complete))
 		assert.Equal(1, job.Count(status.Complete))
 		assert.Equal(1, job.Count(status.Queued))
 		assert.Equal(1, job.Count(status.Errored))
-		assert.Equal(true, job.EndTime.IsZero(), "does not change the original time")
+		assert.Equal(false, job.Running(), "does not change the original time")
 	})
 }